refactor(Product): Value Object 패키지를 분리하라
