+++ src/main/java/kitchenpos/products/tobe/domain/vo/exception/InvalidProductNameException.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain.exception;
+package kitchenpos.products.tobe.domain.vo.exception;
 
 public class InvalidProductNameException extends IllegalArgumentException {
     private static final String MESSAGE = "상품의 이름은 비어있거나, 비속어가 포함될 수 없습니다.";
